Visualization of sole L1 critical factors
In a larger dataset, the files visualize in 3D and 2D the sole factors that are classified as critical for interagency partnership success according to the higher emphasis compared to other factors, posed in the literature (L1 classification). Such visualizations may be used for comparison with the visualizations of the entire dataset.
The entire dataset contains 753 factors and is in the file "Partnership-factor-list-750.xlsx".
The code to draw the visualizations is in the file "Factor_processing_11_revised_KMeans_PCA_diagrams.py".
More information are available on request to nicola.drago@phd.unipd.it.