Create README.md
This repository contains the scripts for downloading, filtering, organizing, and re-naming a project evaluation report and project design report corpus.
These documents constitute a key input for classifying, by criticality, the key factors for predicting inter-agency partnership performance.
The scripts point to co-financed and non-co-financed projects of the Asian Development Bank, the International Fund for Agricultural Development, the United Nations Development Program, and the World Bank.

The scripts were prepared by Mr. Gonçalo Braz of the University of Minho (PT) in the framework and under the specifications and the coordination of the PhD research project of Mr. Nicola Drago of the University of Padova (IT).